% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{average_toxicity}
\alias{average_toxicity}
\title{Average Toxicity Number}
\usage{
average_toxicity(dose, dose_set, true_prob, theta)
}
\arguments{
\item{dose}{a numerical matrix of assigned doses for each step of the trial (column)
and for each trial (row).}

\item{dose_set}{a numerical vector of allowable doses in the trial.}

\item{true_prob}{a numerical vector of the true probabilities associated
with 'dose_set'.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}
}
\value{
Average Toxicity Number.
}
\description{
Calculate the Average Toxicity Number.
}
\references{
Cheung, Y. K. (2011). Dose finding by the continual reassessment method. CRC Press.
}
