% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{avg_perc_dose_allocation}
\alias{avg_perc_dose_allocation}
\title{Average Percentage of Dose Allocation}
\usage{
avg_perc_dose_allocation(sim)
}
\arguments{
\item{sim_list}{an 'ewoc_simulation' object created using the \code{\link[ewoc]{ewoc_simulation}} function.}
}
\value{
Average Percentage of Dose Allocation.
}
\description{
Calculate the Average Percentage of Dose Allocation.
}
