% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlt_curve.R
\name{dlt_curve_d1extended}
\alias{dlt_curve_d1extended}
\title{Plot the DLT curve based on the EWOC extended model}
\usage{
dlt_curve_d1extended(rho, theta, min_dose, max_dose, dose_set = NULL)
}
\arguments{
\item{rho}{a numerical vector indicating the true value of the parameters
rho_0 and rho_1.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}

\item{dose_set}{a numerical vector of doses that probability of DLT will
be calculated.}
}
\description{
Plot the DLT curve based on the EWOC extended model
}
