% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inv_standard_dose}
\alias{inv_standard_dose}
\title{Rescaling standardized dose}
\usage{
inv_standard_dose(dose, min_dose, max_dose)
}
\arguments{
\item{dose}{A numerical value indicating the standardized dose to be
rescaled between \code{min_dose} and \code{max_dose}.}

\item{min_dose}{A numerical value indicating the minimum dose.}

\item{max_dose}{A numerical value indicating the maximum dose.}
}
\value{
A reescaled dose between \code{min_dose} and \code{max_dose}.
}
\description{
Rescaling standardized dose
}
