% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logit}
\alias{logit}
\title{Transform a probability into logit scale}
\usage{
logit(p)
}
\arguments{
\item{p}{A numerical value indicating the probability.}
}
\value{
A numerical value of the logit transformation calculated based on
\code{p}.
}
\description{
Transform a probability into logit scale
}
