% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{mtd_bias}
\alias{mtd_bias}
\title{Bias of the MTD estimates}
\usage{
mtd_bias(mtd_estimate, true_mtd)
}
\arguments{
\item{mtd_estimate}{a numerical vector of the MTD estimates.}

\item{true_mtd}{a numerical value of the true Maximum Tolerable Dose.}
}
\value{
Bias of the MTD estimates.
}
\description{
Calculate the bias.
}
