% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{mtd_selection}
\alias{mtd_selection}
\title{Percentage of MTD Selection}
\usage{
mtd_selection(sim)
}
\arguments{
\item{sim}{an 'ewoc_simulation' object created using the \code{\link[ewoc]{ewoc_simulation}} function.}
}
\value{
Percentage of MTD Selection.
}
\description{
Calculate the Percentage of MTD Selectionn.
}
