% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{optimal_mtd}
\alias{optimal_mtd}
\title{Percent of doses in relation the optimal MTD interval}
\usage{
optimal_mtd(dose_matrix, true_mtd, margin, digits = 2)
}
\arguments{
\item{dose_matrix}{a numerical matrix of assigned doses for each step of the trial (column)
and for each trial (row).}

\item{margin}{a numerical value of the acceptable margin of distance from the
\code{true_MTD}.}

\item{true_MTD}{a numerical value of the true Maximum Tolerable Dose.}
}
\value{
\code{interval} the average percent of doses which are inside the optimal MTD interval.

\code{underdose} the average percent of doses which are smaller than the lower limit of the optimal MTD interval.

\code{overdose} the average percent of doses which are greater than the upper limit of the optimal MTD interval.
}
\description{
Calculate the percent of dose which are inside the optimal MTD interval \code{[true_MTD -
margin ; true_MTD + margin]}.
}
