% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{optimal_mtd_interval}
\alias{optimal_mtd_interval}
\title{Percent of optimal doses}
\usage{
optimal_mtd_interval(dose_matrix, true_mtd, margin)
}
\arguments{
\item{dose_matrix}{a numerical matrix of assigned doses for each step of the trial (column)
and for each trial (row).}

\item{margin}{a numerical value of the acceptable margin of distance from the
\code{true_MTD}.}

\item{true_MTD}{a numerical value of the true Maximum Tolerable Dose.}
}
\value{
\code{trial} the percent of optimal doses for each trial.

\code{average} the average percent of optimal doses.
}
\description{
Calculate the percent of dose which are inside the interval \code{[true_MTD -
margin ; true_MTD + margin]}.
}

