% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{optimal_toxicity}
\alias{optimal_toxicity}
\title{Percent of doses in relation the optimal toxicity interval}
\usage{
optimal_toxicity(dose_matrix, theta, margin, pdlt, digits = 2)
}
\arguments{
\item{dose_matrix}{a numerical matrix of assigned doses for each step of the trial (column)
and for each trial (row).}

\item{theta}{a numerical value of the target DLT rate.}

\item{margin}{a numerical value of the acceptable margin of distance from the
\code{target_rate}.}

\item{pdlt}{a function to calculate the probability of toxicity with a numeric vector of doses as input and a numeric vector of probabilities as output.}

\item{digits}{a numerical value indicating the number of digits.}
}
\value{
\code{interval} the average percent of doses which are inside the optimal toxicity interval.

\code{underdose} the average percent of doses which are smaller than the lower limit of the optimal toxicity interval.

\code{overdose} the average percent of doses which are greater than the upper limit of the optimal toxicity interval.
}
\description{
Calculate the percent of dose which are inside the optimal toxicity interval \code{[target rate -
margin ; target rate + margin]}.
}
