% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{overdose_loss}
\alias{overdose_loss}
\title{Overdose loss function}
\usage{
overdose_loss(mtd_estimate, true_mtd, alpha)
}
\arguments{
\item{mtd_estimate}{a numerical value of the MTD estimate.}

\item{true_mtd}{a numerical value of the true MTD.}

\item{alpha}{Value of the loss for an underdose while the loss for
an overdose is (1 - \code{alpha}).}
}
\value{
value of the evaluated loss function.
}
\description{
Calculate the overdose loss function given by \eqn{alpha(true_mtd -
mtd_estimate)} if \eqn{true_mtd > mtd_estimate} and \eqn{(1 - alpha)
(mtd_estimate - true_mtd)} if \eqn{true_mtd < mtd_estimate}.
}
