% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{overdose_percent}
\alias{overdose_percent}
\title{Percent of dose greater than an upper bound}
\usage{
overdose_percent(dose_matrix, true_mtd, margin = NULL, digits = 2)
}
\arguments{
\item{margin}{a numerical value of the acceptable upper margin of distance from the
\code{true_MTD}.}

\item{dose}{a numerical matrix of assigned doses for each step of the trial (column)
and for each trial (row).}

\item{true_MTD}{a numerical value of the true Maximum Tolerable Dose.}
}
\value{
\code{percent} the percent of the dose assigned which are greater than
\code{true_MTD + margin} for each trial.

\code{average} the average of the percent the dose assigned which are
greater than \code{true_MTD + margin}.
}
\description{
Calculate the percent of dose which are greater than \code{true_mtd + margin}.
}

