% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdlt.R
\name{pdlt_d1dicov}
\alias{pdlt_d1dicov}
\title{Generating a probability of DLT function based on the EWOC with dichotomous covariate model}
\usage{
pdlt_d1dicov(rho, mtd, theta, min_dose, max_dose, levels_cov)
}
\arguments{
\item{rho}{a numerical value indicating the true value of the parameter rho.}

\item{mtd}{a numerical vector indicating the true values of the parameters mtd associated with \code{level_cov}.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}

\item{min_dose}{either a numerical value or a function of the covariate
defining the lower bound of the support of the MTD used to standardize the doses.}

\item{max_dose}{either a numerical value or a function of the covariate
defining the upper bound of the support of the MTD used to standardize the doses.}

\item{levels_cov}{a character vector of the possible values for the covariate.}
}
\value{
A function with dose as an input and a probability based on the
logistic regression and parameters as an output.
}
\description{
Generating a probability of DLT function based on the EWOC with dichotomous covariate model
}
\examples{
pdlt <- pdlt_d1dicov(rho = 0.05, mtd = 60, theta = 0.33,
                    min_dose = 20, max_dose = 100,
                    levels_cov = c("A", "B"))

pdlt(20)

}
