% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response_d1basic}
\alias{response_d1basic}
\title{Generating a response function based on the EWOC basic model}
\usage{
response_d1basic(rho, mtd, theta, min_dose, max_dose)
}
\arguments{
\item{rho}{a numerical value indicating the true value of the parameter rho.}

\item{mtd}{a numerical value indicating the true value of the parameter mtd.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}
}
\description{
Generating a response function based on the EWOC basic model
}
