% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response_d1continuous}
\alias{response_d1continuous}
\title{Generating a binary response function based on the EWOC model with a
continuous covariable}
\usage{
response_d1continuous(rho, mtd, theta, direction, min_dose, max_dose, min_cov,
  max_cov)
}
\arguments{
\item{rho}{a numerical value indicating the true value of the parameter rho.}

\item{mtd}{a numerical value indicating the true value of the parameter mtd
when the covariable is equal to \code{max_cov}.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}

\item{min_cov}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_cov}{a numerical value defining the upper bound of the support of
the MTD.}
}
\description{
Generating a binary response function based on the EWOC model with a
continuous covariable
}
