% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standard_dose}
\alias{standard_dose}
\title{Standardizing dose}
\usage{
standard_dose(dose, min_dose, max_dose)
}
\arguments{
\item{dose}{A numerical value indicating the dose to be standardized between
\code{min_dose} and \code{max_dose}.}

\item{min_dose}{A numerical value indicating the minimum dose.}

\item{max_dose}{A numerical value indicating the maximum dose.}
}
\value{
A standardized dose between 0 and 1.
}
\description{
Standardizing dose
}
