% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_rule.R
\name{stop_rule_sim}
\alias{stop_rule_sim}
\title{Generating a stop rule function}
\usage{
stop_rule_sim(step)
}
\arguments{
\item{step}{an object from the classes 'ewoc_d1classic', 'ewoc_d1extended',
'ewoc_d1ph' generated during the simulation.}
}
\value{
a logical character indicating if the trial should be stopped or not.
}
\description{
Generating a stop rule function
}
\details{
The stop rule function is evaluated at each step of the trial.
It can defined based on any information contained in the object 'step'.
The object 'step' is the output from one of the functions 'ewoc_d1classic',
'ewoc_d1extended', 'ewoc_d1ph'. In this way, a dummy step can be created with
a specific function in order to explore the information available.
}
