% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{trial_simulation.d1extended}
\alias{trial_simulation.d1extended}
\title{Simulation of trials using extended EWOC}
\usage{
\method{trial_simulation}{d1extended}(step_zero, n_sim, sample_size,
  alpha_strategy = c("fixed", "increasing", "conditional"),
  alpha_rate = 0.05, response_sim = NULL, stop_rule_sim = NULL)
}
\arguments{
\item{step_zero}{an object from the class 'ewoc_d1extended'
created using dummy data.}

\item{n_sim}{a number indicating the number of phase I clinical trials
to be simulated.}

\item{sample_size}{a number indicating the number of patients enrolled for
each clinical trial.}

\item{alpha_strategy}{a character indicating the strategy to apply for the
feasibility value. Default is "constant". Options are "increasing" and
"conditional".}

\item{alpha_rate}{a numerical value indicating the rate of the
feasibility strategy. Only necessary if alpha_strategy is either
'increasing' or 'conditional'.}

\item{response_sim}{a function which is self-contained and will be used
as a generator function of the response variables in the simulation.
Its only imput is 'dose' and output is the indicator of DLT.}

\item{stop_rule_sim}{a function having as an imput an object containing all
the information related to the trial as the returned object trial from
\code{ewoc_d1extended} and as output a logical valuel indicating the trial
should be stopped.}
}
\description{
Performing a simulation of several phase I clinical trials based on the extended EWOC.
}
