% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{trial_simulation}
\alias{trial_simulation}
\title{Generic EWOC simulation}
\usage{
trial_simulation(step_zero, n_sim, sample_size, alpha_strategy = "fixed",
  alpha_rate = NULL, response_sim, stop_rule_sim = NULL)
}
\arguments{
\item{step_zero}{an object from the classes 'ewoc_d1classic', 'ewoc_d1extended',
'ewoc_d1ph' created using dummy data.}

\item{n_sim}{a number indicating the number of phase I clinical trials
to be simulated.}

\item{sample_size}{a number indicating the number of patients enrolled for
each clinical trial.}

\item{alpha_strategy}{a character indicating the strategy to apply for the
feasibility value. Default is "constant". Options are "increasing" and
"conditional".}

\item{alpha_rate}{a numerical value indicating the rate of the
feasibility strategy. Only necessary if alpha_strategy is either
'increasing' or 'conditional'.}

\item{response_sim}{a function which is self-contained and will be used
as a generator function of the response variables in the simulation.
Its only imput is 'dose' and output is the indicator of DLT for classical and
extended EWOC and the time until DLT for PH EWOC.}

\item{stop_rule_sim}{a function having as an imput an object containing all
the information related to the trial as the returned object trial from either
\code{ewoc_d1classic}, \code{ewoc_d1extended}, \code{ewoc_d1ph} and as
output a logical valuel indicating the trial should be stopped.}
}
\description{
Generic function for simulating EWOC trials.
}
