% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{dlt_rate}
\alias{dlt_rate}
\title{Evaluation of the DLT rate}
\usage{
dlt_rate(dlt_matrix, trial = FALSE, target_rate = NULL, margin = NULL,
  digits = 2)
}
\arguments{
\item{dlt_matrix}{a matrix of the number of DLT for each step of the trial (column)
and for each trial (row).}

\item{trial}{a logical value indicating if the DLT rate for each trial should be returned.}

\item{target_rate}{a numerical value of the target rate of DLT.}

\item{margin}{a numerical value of the acceptable distance from the \code{target_rate}.}
}
\value{
\code{trial} a numerical vector of the DLT rate for each trial.

\code{average} a numerical value of the average of DLT rate considering a batch of trials.

\code{upper} the percent of trials which the
\code{DLT rate > target_rate + margin} if \code{margin != NULL} and
\code{target_rate != NULL}.

\code{lower} the percent of trials which the
\code{DLT rate < target_rate - margin} if \code{margin != NULL} and
\code{target_rate != NULL}.

\code{interval} the percent of trials which the
\code{target_rate - margin < DLT rate < target_rate + margin} if \code{margin != NULL} and
\code{target_rate != NULL}.
}
\description{
Calculate the DLT rate for each trial, the average DLT rate, the percent
of trials which have \eqn{DLT rate > target_rate + margin}, the percent
of trials which have \eqn{DLT rate < target_rate - margin} and the percent
of trials which have \eqn{target_rate - margin < DLT rate < target_rate + margin}.
}
