% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_EWOC.R
\name{ewoc_d1continuous}
\alias{ewoc_d1continuous}
\title{Escalation Over with Dose Control}
\usage{
ewoc_d1continuous(formula, theta, alpha, mtd_prior, rho_prior, min_dose,
  max_dose, min_cov, max_cov, next_patient_cov, direction = c("positive",
  "negative"), type = c("continuous", "discrete"), first_dose = NULL,
  last_dose = NULL, dose_set = NULL, rounding = c("down", "nearest"),
  n_adapt = 5000, burn_in = 1000, n_mcmc = 1000, n_thin = 1,
  n_chains = 1)
}
\arguments{
\item{formula}{an object of class \code{\link[Formula]{Formula}}: a symbolic
description of the model to be fitted with two regressor parts separated by `|`
corresponding to the dose and covariable, respectively, for the right side and
a matrix as a response containing number of DLT and number of patients for
the left side.}

\item{theta}{a numerical value defining the proportion of expected patients
to experience a medically unacceptable, dose-limiting toxicity (DLT) if
administered the MTD.}

\item{alpha}{a numerical value defining the probability that the dose selected
by EWOC is higher than the MTD.}

\item{mtd_prior}{a matrix of hyperparameters for the Beta prior distribution
associated with the MTD. Each row corresponds to a paramater.}

\item{rho_prior}{a matrix of hyperparameters for the Beta prior distribution
associated with each rho. Each row corresponds to a paramater.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}

\item{next_patient_cov}{a character value indicating the covariable associated to
next patient.}

\item{type}{a character describing the type of the Maximum Tolerable Dose
(MTD) variable.}

\item{first_dose}{a numerical value for the first allowable dose in the trial.
It is only necessary if type = 'continuous'.}

\item{last_dose}{a numerical value for the last allowable dose in the trial.
It is only necessary if type = 'continuous'.}

\item{dose_set}{a numerical vector of allowable doses in the trial. It is only
necessary if type = 'discrete'.}

\item{rounding}{a character indicating how to round a continuous dose to the
one of elements of the dose set. It is only necessary if type = 'discrete'.}

\item{n_adapt}{the number of iterations for adaptation.
See \code{\link[rjags]{adapt}} for details.}

\item{burn_in}{the number of iterations before to start monitoring.}

\item{n_mcmc}{the number of iterations to monitor.}

\item{n_thin}{thinning interval for monitors.}

\item{n_chains}{the number of parallel chains for the model.}

\item{levels_cov}{a character vector of the possible values for the ordinal covariable.}
}
\value{
\code{next_dose} a numerical value corresponding to the next recommend dose.

\code{hpd_dose} a numerical vector containing  the 95\% HPD for the next dose.

\code{pdlt} a numerical value corresponding to the probability of DLT for the next dose.

\code{hpd_pdlt} a numerical vector containing  the 95\% HPD for the probability of DLT for the next dose.

\code{mtd} a numerical vector for the posterior MTD distribution considering the next patient covariable.

\code{rho} a matrix for the posterior rho_00 and rho_01 distributions.

\code{gamma} a numerical vector for the posterior standardized MTD distribution considering the maximum value of the covariable.

\code{sample} a list of the MCMC chains distribution.

\code{trial} a list of trial conditions.
}
\description{
Finding the next dose for a phase I clinical trial based on Escalation Over
Dose Control (EWOC) design considering extended parametrization
for binary response with continuous covariable.
}
\references{
Babb JS, Rogatko A. Patient specific dosing in a cancer phase I clinical trial. Statistics in medicine. 2001 Jul 30;20(14):2079-90.
}

