% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inv_standard_dose}
\alias{inv_standard_dose}
\title{Inverse standardization of the dose}
\usage{
inv_standard_dose(dose, min_dose, max_dose)
}
\arguments{
\item{dose}{a numerical value defining the standardized dose to be unstandardized.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}
}
\value{
\code{dose} a numerical value between \code{min_dose} and \code{max_dose}.
}
\description{
Unstandardizing a dose between minimum and maximum doses.
}
