% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(p)
}
\arguments{
\item{p}{a numerical value defining the probability to be transformed into logit scale.}
}
\value{
\code{logit} a numerical value in logit scale.
}
\description{
Calculating the logit of a probability
}
