% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standard_dose}
\alias{standard_dose}
\title{Standardization of the dose}
\usage{
standard_dose(dose, min_dose, max_dose)
}
\arguments{
\item{dose}{a numerical value defining the dose to be standardized.}

\item{min_dose}{a numerical value defining the lower bound of the support of
the MTD.}

\item{max_dose}{a numerical value defining the upper bound of the support of
the MTD.}
}
\value{
\code{standardized dose} a numerical value between 0 and 1.
}
\description{
Standardizing a dose between 0 and 1.
}
