% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operating_characteristics.R
\name{stop_rule}
\alias{stop_rule}
\title{Evaluation of the stop rule}
\usage{
stop_rule(dlt_matrix, sample_size, digits = 2)
}
\arguments{
\item{dose}{Matrix of the number of DLT for each step of the trial (column)
and for each trial (row).}
}
\value{
A list consisting of
\itemize{
\item{\code{average}: }{Average number of patients to stop a trial.}
\item{\code{min}: }{Minimum number of patients to stop a trial.}
\item{\code{max}: }{Maximum number of patients to stop a trial.}
\item{\code{nstop}: }{Percent of stopped trials}.
}
}
\description{
Calculate the average, minimum, maximum number of patients to stop a trial and
the percent of stopped trials. Stopped trials contain NA after the last
assigned dose.
}

