% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trial_simulation}
\alias{trial_simulation}
\title{Simulation of a trial using Escalation Over with Dose Control}
\usage{
trial_simulation(step0, n_sim, sample_size, alpha_strategy = "fixed",
  alpha_rate = NULL, response_sim)
}
\arguments{
\item{n_sim}{a number indicating the number of phase I clinical trials
to be simulated.}

\item{sample_size}{a number indicating the number of patients enrolled for
each clinical trial.}

\item{alpha_strategy}{a character indicating the strategy to apply for the
feasibility value. Default is "fixed". Options are "increasing" and
"conditional".}

\item{alpha_rate}{a numerical value indicating the rate of the
feasibility strategy. Only necessary if alpha_strategy is either
'increasing' or 'conditional'.}

\item{response_sim}{a function which is self-contained and will beused
as a generator function of the response variables in the simulation.
Its only imput is 'dose' and output is in the same format of the
response variable used to create object 'step_zero'.}

\item{step_zero}{an object from the classes 'ewoc_d1basic', 'ewoc_d1extended',
'ewoc_d1ph', 'ewoc_d1multinomial', 'ewoc_d1ordinal', and 'ewoc_d1continuous'.}
}
\description{
Performing a simulation of several phase I clinical trial based on the
Escalation Over Dose Control (EWOC).
}

