% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{trial_simulation}
\alias{trial_simulation}
\title{Simulation of a trial using Escalation Over with Dose Control}
\usage{
trial_simulation(step_zero, n_sim, sample_size, alpha_strategy = "fixed",
  alpha_rate = NULL, response_sim, stop_rule_sim = NULL)
}
\arguments{
\item{step_zero}{an object from the classes 'ewoc_d1basic', 'ewoc_d1extended',
'ewoc_d1ph', 'ewoc_d1multinomial', 'ewoc_d1ordinal', and 'ewoc_d1continuous'
created using dummy data.}

\item{n_sim}{a number indicating the number of phase I clinical trials
to be simulated.}

\item{sample_size}{a number indicating the number of patients enrolled for
each clinical trial.}

\item{alpha_strategy}{a character indicating the strategy to apply for the
feasibility value. Default is "constant". Options are "increasing" and
"conditional".}

\item{alpha_rate}{a numerical value indicating the rate of the
feasibility strategy. Only necessary if alpha_strategy is either
'increasing' or 'conditional'.}

\item{response_sim}{a function which is self-contained and will be used
as a generator function of the response variables in the simulation.
Its only imput is 'dose' and output is the DLT time.}
}
\description{
Performing a simulation of several phase I clinical trial based on the
Escalation Over Dose Control (EWOC).
}
\examples{
\dontrun{
DLT <- 0
npatients <- 1
dose <- 30
step_zero <- ewoc_d1basic(cbind(DLT, npatients) ~ dose, type = 'discrete',
                         theta = 0.33, alpha = 0.25,
                         min_dose = 0, max_dose = 100,
                         dose_set = seq(0, 100, 20),
                         rho_prior = matrix(1, ncol = 2, nrow = 1),
                         mtd_prior = matrix(1, ncol = 2, nrow = 1),
                         rounding = "nearest")
response_sim <- response_d1basic(rho = 0.05, mtd = 20, theta = 0.33,
                                min_dose = 10, max_dose = 50)
sim <- trial_simulation(step_zero = step_zero,
                       n_sim = 1, sample_size = 30,
                       alpha_strategy = "constant",
                       response_sim = response_sim)

### PH EWOC
time <- 0
status <- 0
dose <- 30

step_zero <- ewoc_d1ph(cbind(time, status) ~ dose, type = 'discrete',
                      theta = 0.33, alpha = 0.25, tau = 10,
                      min_dose = 30, max_dose = 50,
                      dose_set = seq(30, 50, 5),
                      rho_prior = matrix(1, ncol = 2, nrow = 1),
                      mtd_prior = matrix(1, ncol = 2, nrow = 1),
                      distribution = 'exponential',
                      rounding = 'nearest')
response_sim <- response_d1ph(rho = 0.05, mtd = 40, theta = 0.33,
                                min_dose = 30, max_dose = 50,
                                tau = 10, distribution = "exponential")
sim <- trial_simulation(step_zero = step_zero,
                       n_sim = 1, sample_size = 30,
                       alpha_strategy = "increasing",
                       response_sim = response_sim)

}


}

